requirements.txt
scikit-learn~=1.3.0 # because of sklearn
pandas
flask
numpy==1.26.4 # because of evidently
gunicorn
jupyter
ipykernel
evidently
hyperopt
mlflow
psycopg2-binary
azure-identity
azure-storage-blob
seaborn
isort
black
pylint
pytest
pre-commit
python-dotenv
ucimlrepo
